+++ .idea/libraries/animated_vector_drawable_25_0_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="animated-vector-drawable-25.0.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/fe70f53f12fa1829ae1e4abdcf83db8ea27b6bcc/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/fe70f53f12fa1829ae1e4abdcf83db8ea27b6bcc/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/da13157a26af9c4324a09e5518bf040a4c1739d8/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/da13157a26af9c4324a09e5518bf040a4c1739d8/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
