+++ .idea/libraries/appcompat_v7_25_0_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="appcompat-v7-25.0.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/a74979a7d9f80b8b8a506b86a708bbef092e3270/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/a74979a7d9f80b8b8a506b86a708bbef092e3270/output/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/df9ea383e797dcdb3de3d958d2e8d0da24e162c9/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/df9ea383e797dcdb3de3d958d2e8d0da24e162c9/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
