+++ .idea/libraries/design_25_0_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="design-25.0.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/95ef95060d8a8d2b3fe0f2f75ecb5006060f6d38/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/95ef95060d8a8d2b3fe0f2f75ecb5006060f6d38/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/a8c1b62b48fb1fbeadfa748456a8a7a3658d07f2/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/a8c1b62b48fb1fbeadfa748456a8a7a3658d07f2/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
