+++ .idea/libraries/espresso_idling_resource_2_2_2.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="espresso-idling-resource-2.2.2">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/1d7029cf691370e19aadb4777513272a749a0f8d/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/1d7029cf691370e19aadb4777513272a749a0f8d/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/52b9a876889dc3745591a50f8c231fc0a5079e53/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/52b9a876889dc3745591a50f8c231fc0a5079e53/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
