+++ .idea/libraries/fabric_1_3_16.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="fabric-1.3.16">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/ff14f793c96af3561b137aac7d9218db1e72cdbb/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/ff14f793c96af3561b137aac7d9218db1e72cdbb/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/2748daed25c50069b85ba5d369aa48ed33f9210b/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/2748daed25c50069b85ba5d369aa48ed33f9210b/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
