+++ .idea/libraries/facebook_android_sdk_4_20_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="facebook-android-sdk-4.20.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/b8e063defae714df2579ab54d56efe718f62177e/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/b8e063defae714df2579ab54d56efe718f62177e/output/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/d21f9563136e98656ba92ac11fc7cd49d54c9542/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/d21f9563136e98656ba92ac11fc7cd49d54c9542/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
