+++ .idea/libraries/recyclerview_v7_25_0_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="recyclerview-v7-25.0.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/5aeee086b9d930004f0bc3c1858360cd704a018e/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/5aeee086b9d930004f0bc3c1858360cd704a018e/output/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/8097560422e29989968a9e9e70d681c324438bc9/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/8097560422e29989968a9e9e70d681c324438bc9/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
