+++ .idea/libraries/rules_0_5.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="rules-0.5">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/3b530a5880be3f43066655c1df573c35843d4db5/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/3b530a5880be3f43066655c1df573c35843d4db5/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/6509f93f3416e235226c8a4053b758929072315f/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/6509f93f3416e235226c8a4053b758929072315f/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
