+++ .idea/libraries/support_annotations_25_0_0.xml
@@ -1,7 +1,7 @@
 <component name="libraryTable">
   <library name="support-annotations-25.0.0">
     <CLASSES>
-      <root url="jar://C:/Android/Sdk/extras/android/m2repository/com/android/support/support-annotations/25.0.0/support-annotations-25.0.0.jar!/" />
+      <root url="jar://$USER_HOME$/android-sdks/extras/android/m2repository/com/android/support/support-annotations/25.0.0/support-annotations-25.0.0.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
