+++ .idea/libraries/support_compat_25_0_0.xml
@@ -1,9 +1,9 @@
 <component name="libraryTable">
   <library name="support-compat-25.0.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/ac70857d7453b664ab04f33db87c9644eaf60847/output/jars/libs/internal_impl-25.0.0.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/ac70857d7453b664ab04f33db87c9644eaf60847/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/ac70857d7453b664ab04f33db87c9644eaf60847/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/3cf43d703135684705383c22e39c672e980a6313/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/3cf43d703135684705383c22e39c672e980a6313/output/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/3cf43d703135684705383c22e39c672e980a6313/output/jars/libs/internal_impl-25.0.0.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
