+++ .idea/libraries/support_core_ui_25_0_0.xml
@@ -1,9 +1,9 @@
 <component name="libraryTable">
   <library name="support-core-ui-25.0.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/c8a43950ab140dfe37c7db8a0adc8589368cebe4/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/c8a43950ab140dfe37c7db8a0adc8589368cebe4/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/c8a43950ab140dfe37c7db8a0adc8589368cebe4/output/jars/libs/internal_impl-25.0.0.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/5c2b7880761ba9dc2e927de374d7c351182ec39d/output/jars/libs/internal_impl-25.0.0.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/5c2b7880761ba9dc2e927de374d7c351182ec39d/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/5c2b7880761ba9dc2e927de374d7c351182ec39d/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
