+++ .idea/libraries/support_core_utils_25_0_0.xml
@@ -1,9 +1,9 @@
 <component name="libraryTable">
   <library name="support-core-utils-25.0.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/91e297fb7e51d15c63423c88ace2deb15e434e25/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/91e297fb7e51d15c63423c88ace2deb15e434e25/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/91e297fb7e51d15c63423c88ace2deb15e434e25/output/jars/libs/internal_impl-25.0.0.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/0b26d2750d08e2c44de58124a95c6cf5e69ad9f0/output/jars/libs/internal_impl-25.0.0.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/0b26d2750d08e2c44de58124a95c6cf5e69ad9f0/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/0b26d2750d08e2c44de58124a95c6cf5e69ad9f0/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
