+++ .idea/libraries/support_fragment_25_0_0.xml
@@ -1,9 +1,9 @@
 <component name="libraryTable">
   <library name="support-fragment-25.0.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/e63a893281a61f4c4d407ac6e95b0ead6665e556/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/e63a893281a61f4c4d407ac6e95b0ead6665e556/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/e63a893281a61f4c4d407ac6e95b0ead6665e556/output/jars/libs/internal_impl-25.0.0.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/77072474cd96575194a7b3b5868032a895e04062/output/jars/libs/internal_impl-25.0.0.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/77072474cd96575194a7b3b5868032a895e04062/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/77072474cd96575194a7b3b5868032a895e04062/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
