+++ .idea/libraries/support_media_compat_25_0_0.xml
@@ -1,9 +1,9 @@
 <component name="libraryTable">
   <library name="support-media-compat-25.0.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/9faeb07948e49afd38ac1e6d73157a98591b401d/output/jars/classes.jar!/" />
-      <root url="jar://$USER_HOME$/.android/build-cache/9faeb07948e49afd38ac1e6d73157a98591b401d/output/jars/libs/internal_impl-25.0.0.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/9faeb07948e49afd38ac1e6d73157a98591b401d/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/acdbd0e3830a6ca5c30fe3278bffbb0448299339/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/acdbd0e3830a6ca5c30fe3278bffbb0448299339/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/acdbd0e3830a6ca5c30fe3278bffbb0448299339/output/jars/libs/internal_impl-25.0.0.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
