+++ .idea/libraries/support_v4_25_0_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="support-v4-25.0.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/6df6a5ca433ef3f9c1b6bc74ec45300db722b42c/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/6df6a5ca433ef3f9c1b6bc74ec45300db722b42c/output/res" />
+      <root url="file://$USER_HOME$/.android/build-cache/df3176a027d52d2d5f781c16626e203692d6187f/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/df3176a027d52d2d5f781c16626e203692d6187f/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
