+++ .idea/libraries/support_vector_drawable_25_0_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="support-vector-drawable-25.0.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/18eb79c1b0278f20321e0a271083d7dded049e15/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/18eb79c1b0278f20321e0a271083d7dded049e15/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/19bf2758abae08465b183b98a8749697f3b6d68d/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/19bf2758abae08465b183b98a8749697f3b6d68d/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
