+++ .idea/libraries/transition_25_0_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="transition-25.0.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/2337f46d309aff18c5248340b6ddf1805485394e/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/2337f46d309aff18c5248340b6ddf1805485394e/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/46f4c755712e1c669604f8e62e7ae5e4048ded1e/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/46f4c755712e1c669604f8e62e7ae5e4048ded1e/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
