+++ .idea/libraries/twitter_2_3_2.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="twitter-2.3.2">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/4ef7c2c498bfd903d11c8aedc3daa117257e8c33/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/4ef7c2c498bfd903d11c8aedc3daa117257e8c33/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/fb34a971b1a7cf36ff998bf3efd9e3a5432eba0c/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/fb34a971b1a7cf36ff998bf3efd9e3a5432eba0c/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
