+++ .idea/libraries/twitter_core_2_3_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="twitter-core-2.3.1">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/fa8f6069eb4b383a753d4b621782a146695444a8/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/fa8f6069eb4b383a753d4b621782a146695444a8/output/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/072f99f12f711cf383f81b10439645dea0b393e5/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/072f99f12f711cf383f81b10439645dea0b393e5/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
