+++ app/src/main/java/edu/upc/pes/wallachange/LoginSystem/CallbackTwitter.java
@@ -2,12 +2,10 @@ package edu.upc.pes.wallachange.LoginSystem;
 
 import android.widget.Toast;
 
-import com.twitter.sdk.android.core.Callback;
 import com.twitter.sdk.android.core.Result;
 import com.twitter.sdk.android.core.TwitterAuthToken;
 import com.twitter.sdk.android.core.TwitterException;
 import com.twitter.sdk.android.core.TwitterSession;
-import com.twitter.sdk.android.core.identity.TwitterAuthClient;
 
 import edu.upc.pes.wallachange.LoginActivity;
 import edu.upc.pes.wallachange.R;
@@ -26,7 +24,7 @@ public class CallbackTwitter extends com.twitter.sdk.android.core.Callback<Twitt
             TwitterSession session = result.data;
             final String name = session.getUserName();
             final TwitterAuthToken token = session.getAuthToken();
-            myActivity.login(session.getUserId(),session.getUserName(),session.getAuthToken().token,session.getAuthToken().secret);
+            myActivity.login(Long.toString(session.getUserId()),session.getUserName());
             /*
             TwitterAuthClient authClient = new TwitterAuthClient();
             authClient.requestEmail(session, new Callback<String>() {
